Connect to middleware for authentication

+++ routes/web.php
@@ -31,35 +31,38 @@ Route::get('/', [ListingController::class, 'index']);
 // });
 
 //Show Form to Create a Listing
-Route::get('/listing/create', [ListingController::class, 'create']);
+Route::get('/listing/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store a new Listing
-Route::post('/listing', [ListingController::class, 'store']);
+Route::post('/listing', [ListingController::class, 'store'])->middleware('auth');
 
 //Show Form to Edit a Listing
-Route::get('/listing/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listing/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update a Listing
-Route::put('/listing/{listing}', [ListingController::class, 'update']);
+Route::put('/listing/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete a Listing
-Route::delete('/listing/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listing/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+//Manage Listings
+Route::get('/listing/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Show a Listing
 Route::get('/listing/{listing}', [ListingController::class, 'show']);
 
 
 //Show Register/Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 //Create a new User
 Route::post('/users', [UserController::class, 'store']);
 
 //Logout
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 //Show Login Form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 //Login User
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
